+++ lib/dentaku/ast/functions/map.rb
@@ -1,23 +1,13 @@
-require_relative '../function'
+require_relative './enum'
 require_relative '../../exceptions'
 
 module Dentaku
   module AST
-    class Map < Function
-      def self.min_param_count
-        3
-      end
-
-      def self.max_param_count
-        3
-      end
-
-      def deferred_args
-        [1, 2]
-      end
-
+    class Map < Enum
       def value(context = {})
-        collection      = @args[0].value(context)
+        validate_identifier(@args[1])
+
+        collection      = Array(@args[0].value(context))
         item_identifier = @args[1].identifier
         expression      = @args[2]